python_server/pyproject.toml
[project]
name = "wingechr-python-server"
version = "0.0.0"

requires-python = ">=3.10"
dependencies = [
  "django>=5,<6",
  "django-compressor>=4.5",
  "django-cors-headers~=4.6.0",
  "django-filter~=24.3",
  "django-htmlmin~=0.11",
  "djangorestframework>=3.10",
  "drf-spectacular[sidecar]~=0.28.0",
  "html5validator~=0.4",
  "psycopg2-binary~=2.9",
]

#!/bin/bash
# Build: create venv and _static

#test -e venv || python3 -m venv venv
#. venv/bin/activate
#pip install -r requirements.txt
## node/npm
#npm install --no-save --omit=dev
# copy default settings
#mkdir -p _local
#cp app/local_settings.example.py _local/local_settings.py
#python app/manage.py collectstatic --noinput
#python app/manage.py compress --force
# TODO:translations python app/manage.py compilemessages

#!/bin/bash
# requires build.sh to run first, requires variable DEPLOY_PATH

#if [[ -z "${DEPLOY_PATH}" ]]; then
#    echo "DEPLOY_PATH is not defined" >&2
#    exit 1
#fi
#mkdir -p "$DEPLOY_PATH"
#
## copy app and build data
#rsync -au --exclude "**/__pycache__" --delete app "$DEPLOY_PATH"
#rsync -au --delete venv "$DEPLOY_PATH"
#rsync -au --delete _static "$DEPLOY_PATH"
#
#pushd "$DEPLOY_PATH"
#. venv/bin/activate
## copy default settings if not exist
#mkdir -p _local
#test -e _local/local_settings.py || cp app/local_settings.example.py _local/local_settings.py
## update deploy system
## NOTE: ignore warnings that node_modules in the STATICFILES_DIRS setting does not exist
## we have copied already all static files
#python app/manage.py migrate
#python app/manage.py create_default_users # does nothing if already exist
#
## restart
#touch app/wsgi.py
#popd
